Break circular dependencies, make tests mirror src
