+++ src/blueapi/client/plan_cache.py
@@ -27,10 +27,7 @@ class PlanFailedError(Exception):
 
 class PlanCache:
     def __init__(self, client: ClientProtocol, plans: list[PlanModel]):
-        self._cache = {
-            model.name: Plan(name=model.name, model=model, client=client)
-            for model in plans
-        }
+        self._cache = {model.name: Plan(model=model, client=client) for model in plans}
         for name, plan in self._cache.items():
             if name.startswith("_"):
                 continue
@@ -50,15 +47,14 @@ class PlanCache:
 
 
 class Plan:
-    def __init__(self, name, model: PlanModel, client: ClientProtocol):
-        self.name = name
+    def __init__(self, model: PlanModel, client: ClientProtocol):
         self.model = model
         self._client = client
         self.__doc__ = model.description
 
     def __call__(self, *args, **kwargs) -> Any:
         req = TaskRequest(
-            name=self.name,
+            name=self.model.name,
             params=self._build_args(*args, **kwargs),
             instrument_session=self._client.instrument_session,
         )
@@ -89,9 +85,9 @@ class Plan:
         )
 
         if len(args) > len(self.properties):
-            raise TypeError(f"{self.name} got too many arguments")
+            raise TypeError(f"{self.model.name} got too many arguments")
         if extra := {k for k in kwargs if k not in self.properties}:
-            raise TypeError(f"{self.name} got unexpected arguments: {extra}")
+            raise TypeError(f"{self.model.name} got unexpected arguments: {extra}")
 
         params = {}
         # Initially fill parameters using positional args assuming the order
@@ -103,7 +99,7 @@ class Plan:
         for key, value in kwargs.items():
             # If we've already assumed a positional arg was this value, bail out
             if key in params:
-                raise TypeError(f"{self.name} got multiple values for {key}")
+                raise TypeError(f"{self.model.name} got multiple values for {key}")
             params[key] = value
 
         if missing := {k for k in self.required if k not in params}:
@@ -124,7 +120,7 @@ class Plan:
             return f"{name}: {typ} | None = None"
 
         args = [_format_arg(name, info) for name, info in self.properties.items()]
-        single_line = f"{self.name}({', '.join(args)})"
+        single_line = f"{self.model.name}({', '.join(args)})"
 
         if len(single_line) <= _REPR_MAX_LENGTH and len(args) <= _REPR_MAX_ARGS_INLINE:
             return single_line
@@ -132,7 +128,7 @@ class Plan:
         indent = "    "
         # Fall back to multiline if too many arguments or too long.
         multiline_args = ",\n".join(f"{indent}{arg}" for arg in args)
-        return f"{self.name}(\n{multiline_args}\n)"
+        return f"{self.model.name}(\n{multiline_args}\n)"
 
 
 def _pretty_type(schema: dict[str, Any]) -> str:
