+++ src/blueapi/client/protocols.py
@@ -1,6 +1,6 @@
 from typing import Protocol
 
-from blueapi.service.model import TaskRequest
+from blueapi.service.model import DeviceModel, DeviceResponse, TaskRequest
 from blueapi.worker.event import TaskStatus
 
 from .event_bus import OnAnyEvent
@@ -16,3 +16,13 @@ class ClientProtocol(Protocol):
 
     @property
     def instrument_session(self) -> str: ...
+
+
+class RestClientProtocol(Protocol):
+    def get_devices(self) -> DeviceResponse: ...
+
+    def get_device(self, name: str) -> DeviceModel: ...
+
+
+class ClientObjectRef:
+    model: DeviceModel
